Specialize client for IOS XR
